requirements.txt
python-dotenv~=1.0.1
unstructured==0.15.6
langchain==0.2.14
chromadb==0.5.5
langchain-text-splitters==0.2.2
langchain-community==0.2.12
langchain-core==0.2.33
bs4==0.0.2
langchain-openai==0.1.22
tiktoken==0.7.0
tqdm==4.66.4
faiss-cpu==1.8.0.post1
requests~=2.32.3
beautifulsoup4~=4.12.3
streamlit==1.37.1
ollama==0.3.1